WIP: component refactoring and dark mode
- Rename AppChrome → AppHeader, IndexView → DirectoryView
- Extract SpeakerFilterBar and YearFilterBar from shared filter UI
- Add AppFooter component to unify footer across all pages
- Add dark mode CSS variables via prefers-color-scheme
